# Improve joiner performance for random-access separators