As mentioned in Jon's talk today, `joiner` is __really slow__.
This small improvement that was laying around for far too long.

```
std.joiner      = 5 secs, 180 ms, 193 μs, and 7 hnsecs
std.joiner.alias = 5 secs, 106 ms, and 2 μs
new.joiner      = 3 secs, 168 ms, 560 μs, and 6 hnsecs
foreach         = 2 secs, 537 ms, 962 μs, and 8 hnsecs
```
(ldc 1.9.0)

Benchmark code: https://gist.github.com/wilzbach/ffd5d20639766a831fd4b1962572897a

FWIW dmd is playing in a different universe here:

```
> dmd -O -inline -release ./joiner.d && ./joiner
std.joiner      = 57 secs, 834 ms, 289 μs, and 3 hnsecs
std.joiner.alias = 54 secs, 836 ms, 908 μs, and 6 hnsecs
new.joiner      = 44 secs, 936 ms, 706 μs, and 5 hnsecs
foreach         = 23 secs, 732 ms, 820 μs, and 3 hnsecs
```