@wilzbach 2018-05-02
BTW I think @jondegenhardt who discovered this might be interested in this.

(The issue is https://issues.dlang.org/buglist.cgi?quicksearch=joiner&list_id=221332, but I didn't reference it here because there's still a quite a difference between the naive loop and joiner.)