@wilzbach 2018-05-31
Thanks for the review and feedback @jondegenhardt!
Finally got around addressing the minor things.