@jondegenhardt 2018-06-17
@wilzbach I don't see this in the 2.081.0 Beta-1 release notes. Seems worthy of inclusion, even if there is more to be gained.