@wilzbach 2018-06-17
It's part of the 2.081 Beta-1 release:

![image](https://user-images.githubusercontent.com/4370550/41506680-b3c34fba-7222-11e8-8abd-1d1c4168788b.png)

However, it doesn't show up in the release notes because I didn't reference [18643](https://issues.dlang.org/show_bug.cgi?id=18463) as fixed by this PR because as you said yourself there's still more to gain.
It probably makes sense to add a special changelog entry anyway as it's "free advertisement" and showing that we slowly fix our issues (will do so tonight).
Thanks a lot for the ping and reminder!